pyproject.toml
[project]
name = "mandelbrot"
version = "0.1.0"
description = "Live deomonstration of the Mandelbrot set to illustrate performance improvements in Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "colorcet>=3.1.0",
    "mandelbrot-implementation-rust-native",
    "mandelbrot-implementations-cython-direct",
    "mandelbrot-implementations-cython-naive",
    "mandelbrot-implementations-numba",
    "numpy>=2.2.6",
    "pygame>=2.6.1",
]

[project.gui-scripts]
mandelbrot = "mandelbrot.__main__:run"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.uv.sources]
mandelbrot-implementations-cython-naive = { path = "src/mandelbrot/implementations/cython/cython_naive" }
mandelbrot-implementations-cython-direct = { path = "src/mandelbrot/implementations/cython/cython_direct" }
mandelbrot-implementations-numba = { path = "src/mandelbrot/implementations/numba/numba_jit" }
mandelbrot-implementation-rust-native = { path = "src/mandelbrot/implementations/rust/rust_native" }


[tool.ruff.lint]
select = ["I"]
